Refactored Socket types
